Thomas Neele reported that under Windows two tests were hanging. This change solves at least one test. It was probably a pointer that got out of range if the first state of the model has no outgoing transitions.
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@15135 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ libraries/lts/include/mcrl2/lts/detail/check_complexity.h
@@ -287,7 +287,7 @@ class check_complexity
                                                          __builtin_clzll(size);
             }
         #endif
-        return std::log2(size);
+        return (int) std::log2(size);
     }
 
   private:
